+++ flake.nix
@@ -1,16 +1,12 @@
-# 💫 https://github.com/JaKooLit 💫 #
-
 {
   description = "KooL's NixOS-Hyprland"; 
   	
   inputs = {
-	nixpkgs.url = "nixpkgs/nixos-24.11";
-  	#nixpkgs.url = "nixpkgs/nixos-unstable";
+	#nixpkgs.url = "nixpkgs/nixos-24.11";
+  	nixpkgs.url = "nixpkgs/nixos-unstable";
 	#hyprland.url = "github:hyprwm/Hyprland"; # hyprland development
-
-	ags.url = "github:aylur/ags/v1"; # aylurs-gtk-shell-v1
-	
-	#distro-grub-themes.url = "github:AdisonCavani/distro-grub-themes"; #for grub themes
+	#distro-grub-themes.url = "github:AdisonCavani/distro-grub-themes";
+	#ags.url = "github:aylur/ags/v1"; # aylurs-gtk-shell-v1
   	};
 
   outputs = 
@@ -38,7 +34,7 @@
 			};
 	   		modules = [ 
 				./hosts/${host}/config.nix 
-				#inputs.distro-grub-themes.nixosModules.${system}.default #for grub themes
+				# inputs.distro-grub-themes.nixosModules.${system}.default
 				];
 			};
 		};