+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelogs
 
+## 15 Feb 2025
+- Returning AGS overview option :)
+
+## 13 Feb 2025
+- disabled AGS permanently as it is trying to install version 2.2.1 even I set override
+- switched to latest kernel (from zen kernel) to unstable channel (from 24.11)
+
 ## 30 Jan 2025
 - AGS (aylurs gtk shell) v1 is now optional